+++ test/unit/services/service_listeners/publishing_api_pusher_test.rb
@@ -9,11 +9,10 @@ module ServiceListeners
         .with(edition.id, event)
     end
 
-    def stub_html_attachment_pusher(edition, event, unpublishing = nil)
-      PublishingApiHtmlAttachmentsWorker
-        .any_instance
-        .expects(:perform)
-        .with(edition.id, event, unpublishing.try(:id))
+    def stub_html_attachment_pusher(edition, event)
+      PublishingApiHtmlAttachments
+        .expects(:process)
+        .with(edition, event)
     end
 
     def stub_publications_pusher(edition, event)
@@ -103,7 +102,7 @@ module ServiceListeners
     test "unpublish publishes the unpublishing" do
       edition = create(:unpublished_publication)
       Whitehall::PublishingApi.expects(:unpublish_async).with(edition.unpublishing)
-      stub_html_attachment_pusher(edition, "unpublish", edition.unpublishing)
+      stub_html_attachment_pusher(edition, "unpublish")
       stub_publications_pusher(edition, "unpublish")
       Sidekiq::Testing.inline! do
         PublishingApiPusher.new(edition).push(event: "unpublish")
@@ -175,7 +174,6 @@ module ServiceListeners
       fr = res[:deleted_translation]
 
       Whitehall::PublishingApi.expects(:publish).with(new_edition)
-      stub_html_attachment_pusher(new_edition, "publish")
 
       pusher = mock
       PublishingApiGoneWorker
