+++ test/unit/workers/publishing_api_document_republishing_worker_test.rb
@@ -22,15 +22,13 @@ class PublishingApiDocumentRepublishingWorkerTest < ActiveSupport::TestCase
       .with(draft_edition, "republish", false)
 
     invocation_order = sequence("invocation_order")
-    PublishingApiHtmlAttachmentsWorker
-      .any_instance
-      .expects(:perform)
-      .with(published_edition.id, "republish")
+    ServiceListeners::PublishingApiHtmlAttachments
+      .expects(:process)
+      .with(published_edition, "republish")
       .in_sequence(invocation_order)
-    PublishingApiHtmlAttachmentsWorker
-      .any_instance
-      .expects(:perform)
-      .with(draft_edition.id, "republish")
+    ServiceListeners::PublishingApiHtmlAttachments
+      .expects(:process)
+      .with(draft_edition, "republish")
       .in_sequence(invocation_order)
 
     PublishingApiDocumentRepublishingWorker.new.perform(document.id)
@@ -109,15 +107,13 @@ class PublishingApiDocumentRepublishingWorkerTest < ActiveSupport::TestCase
     unpublishing_worker.expects(:perform).with(published_edition.unpublishing.id, false)
 
     invocation_order = sequence("invocation_order")
-    PublishingApiHtmlAttachmentsWorker
-      .any_instance
-      .expects(:perform)
-      .with(published_edition.id, "republish")
+    ServiceListeners::PublishingApiHtmlAttachments
+      .expects(:process)
+      .with(published_edition, "republish")
       .in_sequence(invocation_order)
-    PublishingApiHtmlAttachmentsWorker
-      .any_instance
-      .expects(:perform)
-      .with(published_edition.id, "republish")
+    ServiceListeners::PublishingApiHtmlAttachments
+      .expects(:process)
+      .with(published_edition, "republish")
       .in_sequence(invocation_order)
 
     PublishingApiDocumentRepublishingWorker.new.perform(document.id)
